+++ conda_requirements.txt
@@ -1,4 +1,5 @@
 python>=3.4
 pandas
 scipy
+numpy
 nose=1.3.7
